Pull game component out of router navigation/view.
