+++ src/web/src/views/AboutView.vue
@@ -1,15 +1,17 @@
 <template>
-  <p>
-    Play snake with your friends!
-  </p>
+  <div class="m-2">
+    <p>
+      Play snake with your friends!
+    </p>
 
-  <h3>
-    Attributions
-  </h3>
+    <h3>
+      Attributions
+    </h3>
 
-  <p>
-    The snake favicon is used under the <a href="https://creativecommons.org/licenses/by/4.0/">CC-BY 4.0</a> license from the <a href="https://github.com/twitter/twemoji">twemoji</a> project.
-  </p>
+    <p>
+      The snake favicon is used under the <a href="https://creativecommons.org/licenses/by/4.0/">CC-BY 4.0</a> license from the <a href="https://github.com/twitter/twemoji">twemoji</a> project.
+    </p>
+  </div>
 </template>
 
 <style>
